+++ src/printer_model_smartpy.ml
@@ -99,10 +99,10 @@ let pp_model fmt (model : model) =
 
   let pp_api_asset fmt = function
     | Get _ -> ()
-      (* Format.fprintf fmt
-        "def get_%s (self, key):@\n  \
-         self.data.%s_assets[key]@\n"
-        an an *)
+    (* Format.fprintf fmt
+       "def get_%s (self, key):@\n  \
+       self.data.%s_assets[key]@\n"
+       an an *)
 
     | Set an ->
       Format.fprintf fmt
@@ -260,6 +260,9 @@ let pp_model fmt (model : model) =
     | Bconcat t -> Format.fprintf fmt "concat on %a" pp_type t
     | Bslice  t -> Format.fprintf fmt "slice on %a"  pp_type t
     | Blength t -> Format.fprintf fmt "length on %a" pp_type t
+    | Bisnone t -> Format.fprintf fmt "isnone on %a" pp_type t
+    | Bissome t -> Format.fprintf fmt "issome on %a" pp_type t
+    | Bgetopt t -> Format.fprintf fmt "getopt on %a" pp_type t
   in
 
   let pp_api_internal fmt = function
@@ -849,13 +852,13 @@ let pp_model fmt (model : model) =
       (* utils *)
 
       | Mcast (_src, _dst, v) -> f fmt v
-        (* let pp fmt (src, dst, v) =
-          Format.fprintf fmt "cast_%a_%a(%a)"
-            pp_type src
-            pp_type dst
-            f v
-        in
-        pp fmt (src, dst, v) *)
+      (* let pp fmt (src, dst, v) =
+         Format.fprintf fmt "cast_%a_%a(%a)"
+          pp_type src
+          pp_type dst
+          f v
+         in
+         pp fmt (src, dst, v) *)
 
       | Mgetfrommap (an, k, c) ->
         let pp fmt (an, k, c) =
@@ -923,6 +926,17 @@ let pp_model fmt (model : model) =
         Format.fprintf fmt "length (%a)"
           f x
 
+      | Misnone x ->
+        Format.fprintf fmt "isnone (%a)"
+          f x
+
+      | Missome x ->
+        Format.fprintf fmt "issome (%a)"
+          f x
+
+      | Mgetopt x ->
+        Format.fprintf fmt "getopt (%a)"
+          f x
 
       (* crypto functions *)