+++ CHANGES.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Crypto functions
  - `slice`, `concat` and `length` functions
+ - `isnone`, `issome`, `getopt` option functions
 
 ## [0.1.13] - 2020-03-05
 ### Added